support for augmentation data when building .eh_frame. As of this commit, emitting dwarfscript, compiling that dwarfscript, inserting the new section into the original object and then emitting dwarfscript is an identity operation. Therefore, dwarfscript compilation may be deemed to work.
This only works currently when the .eh_frame does not
expand. Currently when the section grows it overflows into the next
section and all manner of things can go wrong with it. Fixing this
problem is the next course of action (although it somewhat tricky
because we don't necessarily have relocation information available)